+++ lib/src/easy_infinite_date_time/widgets/easy_infinite_date_timeline_controller.dart
@@ -23,7 +23,8 @@ final class EasyInfiniteDateTimelineController {
       'EasyInfiniteDateTimelineController is not attached to any EasyInfiniteDateTimeline View.',
     );
 
-    final offset = _calculateDateOffset(_infiniteTimeLineState!._focusDate);
+    final offset = _infiniteTimeLineState!
+        ._getScrollOffset(_infiniteTimeLineState!._focusDate);
     _infiniteTimeLineState!._controller.jumpTo(offset);
   }
 
@@ -38,7 +39,8 @@ final class EasyInfiniteDateTimelineController {
       'EasyInfiniteDateTimelineController is not attached to any EasyInfiniteDateTimeline View.',
     );
 
-    final offset = _calculateDateOffset(_infiniteTimeLineState!._focusDate);
+    final offset = _infiniteTimeLineState!
+        ._getScrollOffset(_infiniteTimeLineState!._focusDate);
     _infiniteTimeLineState!._controller.animateTo(
       offset,
       duration: duration,
@@ -57,7 +59,7 @@ final class EasyInfiniteDateTimelineController {
       _infiniteTimeLineState != null,
       'EasyInfiniteDateTimelineController is not attached to any EasyInfiniteDateTimeline View.',
     );
-    final offset = _calculateDateOffset(date);
+    final offset = _infiniteTimeLineState!._getScrollOffset(date);
     _infiniteTimeLineState!._controller.animateTo(
       offset,
       duration: duration,
@@ -75,7 +77,7 @@ final class EasyInfiniteDateTimelineController {
       _infiniteTimeLineState != null,
       'EasyInfiniteDateTimelineController is not attached to any EasyInfiniteDateTimeline View.',
     );
-    final offset = _calculateDateOffset(DateTime.now());
+    final offset = _infiniteTimeLineState!._getScrollOffset(DateTime.now());
     _infiniteTimeLineState!._controller.animateTo(
       offset,
       duration: duration,
@@ -92,31 +94,9 @@ final class EasyInfiniteDateTimelineController {
       _infiniteTimeLineState != null,
       'EasyInfiniteDateTimelineController is not attached to any EasyInfiniteDateTimeline View.',
     );
-    final offset = _calculateDateOffset(date);
+    final offset = _infiniteTimeLineState!._getScrollOffset(date);
     _infiniteTimeLineState!._controller.jumpTo(
       offset,
     );
-  }
-
-  /// the method calculates the number of days between startDate and lastDate using the difference() method
-  /// of the Duration class. This value is stored in the offset variable.
-  /// If offset is equal to 0, the method returns 0.0.
-  /// Otherwise, the method calculates the horizontal offset of the day
-  /// by multiplying the offset value by the width of a day widget
-  /// (which is either the value of widget.easyDayProps.width or a default value of EasyConstants.dayWidgetWidth).
-  /// It then adds to this value the product of offset and [EasyConstants.separatorPadding] (which represents the width of the space between each day widget)
-  double _calculateDateOffset(DateTime lastDate) {
-    final firstDate = _infiniteTimeLineState!.widget.firstDate;
-    int offset = lastDate.difference(firstDate).inDays;
-    double adjustedHPadding = _infiniteTimeLineState!._timeLineProps.hPadding >
-            EasyConstants.timelinePadding
-        ? (_infiniteTimeLineState!._timeLineProps.hPadding -
-            EasyConstants.timelinePadding)
-        : 0.0;
-    if (offset == 0) {
-      return 0.0;
-    }
-    return (offset * _infiniteTimeLineState!._dayOffsetConstrains) +
-        adjustedHPadding;
   }
 }