+++ lib/src/properties/properties.dart
@@ -2,4 +2,5 @@ export 'date_formatter.dart';
 export 'easy_day_props.dart';
 export 'day_style.dart';
 export 'easy_header_props.dart';
+export 'selection_mode.dart';
 export 'time_line_props.dart';